requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
phidata
pgvector==0.3.6
googlesearch-python==1.3.0
sentence-transformers==3.4.1
psycopg-binary
torch==2.2.2
gunicorn==21.2.0
sqlalchemy==2.0.19
pycountry==22.3.5
groq
numpy<2
psycopg2
huggingface_hub[hf_xet]

